+++ client/data/mapDetails.js
